Fix different types

+++ include/globals.h
@@ -9,5 +9,5 @@ extern const char *credits;
 extern char serial_in;
 extern char serial_out;
 extern volatile char kbc;
-extern int uptime_ticks;
+extern uint64_t uptime_ticks;
 #endif
